Riorganizzate solutions Aggiunta prima prova di unitTest Generator - Rimosse tutte le segnalazioni SonarQube

+++ Generator/Classes/Record/Columns_Record.cs
@@ -12,7 +12,7 @@
         public string DATA_TYPE { get; set; }
         public int? CHARACTER_MAXIMUM_LENGTH { get; set; }
         public int? CHARACTER_OCTET_LENGTH { get; set; }
-        public int? NUMERIC_PRECISION { get; set; }
+        public byte? NUMERIC_PRECISION { get; set; }
         public int? NUMERIC_PRECISION_RADIX { get; set; }
         public int? NUMERIC_SCALE { get; set; }
         public short? DATETIME_PRECISION { get; set; }
